Convert finish login pages to class components. (#2365)
* Convert finish login pages to class components.

* convert to JSX
