+++ packages/commonwealth/client/scripts/views/pages/finish_axie_login.tsx
@@ -1,9 +1,12 @@
-import m from 'mithril';
-import $ from 'jquery';
-import app from 'state';
+/* @jsx m */
 
-import { PageLoading } from 'views/pages/loading';
+import m from 'mithril';
+import ClassComponent from 'class_component';
+import $ from 'jquery';
+
+import app from 'state';
 import { initAppState } from 'app';
+import { PageLoading } from 'views/pages/loading';
 import { updateActiveAddresses } from 'controllers/app/login';
 import ErrorPage from './error';
 
@@ -42,27 +45,34 @@ const validate = async (
   }
 };
 
-const FinishAxieLogin: m.Component<Record<string, unknown>, IState> = {
-  oninit: (vnode) => {
+class FinishAxieLogin extends ClassComponent<Record<string, unknown>> {
+  private state: IState = {
+    validating: false,
+    error: '',
+  };
+
+  public oninit() {
     // grab token
     // TODO: how to use state id?
     const token = m.route.param('token');
     const stateId = m.route.param('stateId');
+
     validate(token, stateId, 'axie-infinity').then((res) => {
       if (typeof res === 'string') {
-        vnode.state.error = res;
+        this.state.error = res;
         m.redraw();
       }
     });
-  },
-  view: (vnode) => {
+  }
+
+  public view() {
     console.log('finish axie login');
-    if (vnode.state.error) {
-      return m(ErrorPage, { title: 'Login Error', message: vnode.state.error });
+    if (this.state.error) {
+      return <ErrorPage title="Login Error" message={this.state.error} />;
     } else {
-      return m(PageLoading);
+      return <PageLoading />;
     }
-  },
-};
+  }
+}
 
 export default FinishAxieLogin;
